+++ src/components/domain/CardCVCForm.tsx
@@ -1,13 +1,11 @@
-import { useEffect, useState } from 'react';
-import isZeroOrExactLength from '../../utils/isExactLength';
 import NumberInput from '../common/NumberInput';
-import { ERROR_MESSAGE } from '../../constants/guide';
 import {
   NumberInputForm,
   Label,
   NumberInputContainer,
   ErrorText,
 } from '../../styles/CardForm.styles';
+import useCardCVCValidation from '../../hooks/Validation/useCardCVCValidation';
 
 interface CardCVCFormProps {
   cardInfo: CardInfo;
@@ -16,15 +14,7 @@ interface CardCVCFormProps {
 }
 
 function CardCVCForm({ cardInfo, handleCardInfo, maxLength }: CardCVCFormProps) {
-  const [errorText, setErrorText] = useState('');
-
-  useEffect(() => {
-    if (isZeroOrExactLength(cardInfo.cvc, maxLength)) {
-      setErrorText('');
-    } else {
-      setErrorText(ERROR_MESSAGE.GET_LENGTH_TEXT(maxLength));
-    }
-  }, [cardInfo.cvc]);
+  const { errorText } = useCardCVCValidation(cardInfo, maxLength);
 
   return (
     <NumberInputForm>
