+++ src/components/domain/CardExpirationForm.tsx
@@ -1,13 +1,11 @@
-import { useEffect, useState } from 'react';
-import isZeroOrExactLength from '../../utils/isExactLength';
 import NumberInput from '../common/NumberInput';
-import { ERROR_MESSAGE } from '../../constants/guide';
 import {
   NumberInputForm,
   Label,
   NumberInputContainer,
   ErrorText,
 } from '../../styles/CardForm.styles';
+import useCardExpirationValidation from '../../hooks/Validation/useCardExpirationValidation';
 
 interface CardExpirationFormProps {
   cardInfo: CardInfo;
@@ -20,46 +18,20 @@ interface CardExpirationFormProps {
 }
 
 function CardExpirationForm({ cardInfo, handleCardInfo, maxLength }: CardExpirationFormProps) {
-  const [errorText, setErrorText] = useState('');
-
-  const isValidMonth =
-    Number(cardInfo.expiration.month) >= 1 && Number(cardInfo.expiration.month) <= 12;
-  const isValidYear =
-    Number(cardInfo.expiration.year) >= 25 && Number(cardInfo.expiration.year) <= 99;
-
-  useEffect(() => {
-    const isExactDigits = [cardInfo.expiration.month, cardInfo.expiration.year].some((number) => {
-      if (isZeroOrExactLength(number, maxLength)) return false;
-      return true;
-    });
-
-    if (isExactDigits) {
-      setErrorText(ERROR_MESSAGE.GET_LENGTH_TEXT(maxLength));
-      return;
-    }
-    if (cardInfo.expiration.month !== '' && !isValidMonth) {
-      setErrorText(ERROR_MESSAGE.INVALID_MONTH);
-      return;
-    }
-    if (cardInfo.expiration.year !== '' && !isValidYear) {
-      setErrorText(ERROR_MESSAGE.INVALID_YEAR);
-      return;
-    }
-    setErrorText('');
-  }, [cardInfo.expiration.month, cardInfo.expiration.year]);
+  const { errorText } = useCardExpirationValidation(cardInfo, maxLength);
 
   const InputInfo = [
     {
       value: cardInfo.expiration.month,
       setValue: (value: string) => handleCardInfo('expiration', value, 'month'),
       placeholder: 'MM',
-      extraErrorCondition: cardInfo.expiration.month !== '' && !isValidMonth,
+      // extraErrorCondition: cardInfo.expiration.month !== '' && !isValidMonth,
     },
     {
       value: cardInfo.expiration.year,
       setValue: (value: string) => handleCardInfo('expiration', value, 'year'),
       placeholder: 'YY',
-      extraErrorCondition: cardInfo.expiration.year !== '' && !isValidYear,
+      // extraErrorCondition: cardInfo.expiration.year !== '' && !isValidYear,
     },
   ];
 
@@ -74,7 +46,7 @@ function CardExpirationForm({ cardInfo, handleCardInfo, maxLength }: CardExpirat
             setValue={inputInfo.setValue}
             maxLength={maxLength}
             placeholder={inputInfo.placeholder}
-            extraErrorCondition={inputInfo.extraErrorCondition}
+            // extraErrorCondition={inputInfo.extraErrorCondition}
           />
         ))}
       </NumberInputContainer>
