+++ src/stories/components/domain/CardCVCForm.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta } from '@storybook/react';
 import CardCVCForm from '../../../components/domain/CardCVCForm';
-import useCardInfo from '../../../hooks/useCardInfo';
+import useCardInfo from '../../../hooks/useCardInfoState';
 
 const meta = {
   title: 'Components/domain/CardCVCForm',
