+++ src/stories/components/domain/CardNumberForm.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta } from '@storybook/react';
 import CardNumberForm from '../../../components/domain/CardNumberForm';
-import useCardInfo from '../../../hooks/useCardInfo';
+import useCardInfo from '../../../hooks/useCardInfoState';
 const meta = {
   title: 'Components/domain/CardNumberForm',
   component: CardNumberForm,
